# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/emergencies/{emergencyId}/resources": {
      "post": {
        "operationId": "ResourcesController_create",
        "parameters": [
          {
            "name": "emergencyId",
            "required": true,
            "in": "path",
            "description": "Emergency UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or UUID"
          },
          "401": {
            "description": "Missing or invalid token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a resource for an emergency (requires authentication)",
        "tags": [
          "resources"
        ]
      }
    },
    "/resources/{resourceId}/verify": {
      "post": {
        "operationId": "ResourcesController_verifyResource",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "description": "Resource UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyResourceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Resource verified"
          },
          "400": {
            "description": "Invalid verification level or UUID"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Coordinator role required"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify a resource (coordinator of the resource's emergency only)",
        "tags": [
          "resources"
        ]
      }
    },
    "/resources/{resourceId}/publish": {
      "post": {
        "operationId": "ResourcesController_publishResource",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "description": "Resource UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Resource published"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Coordinator role required"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource not verified yet"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a resource (coordinator of the resource's emergency only)",
        "tags": [
          "resources"
        ]
      }
    },
    "/resources/{resourceId}": {
      "patch": {
        "operationId": "ResourcesController_editResourceFields",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "description": "Resource UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditResourceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Resource edited"
          },
          "400": {
            "description": "Missing reason or resource is discarded"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Validator/coordinator role required"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a resource during verification (validator/coordinator). Requires a reason; recorded in the audit trail.",
        "tags": [
          "resources"
        ]
      }
    },
    "/resources/{resourceId}/discard": {
      "post": {
        "operationId": "ResourcesController_discardResourceAction",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "description": "Resource UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscardResourceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Resource discarded"
          },
          "400": {
            "description": "Missing reason or resource is not pending verification"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Validator/coordinator role required"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Discard a resource during verification (validator/coordinator). Requires a reason; recorded in the audit trail.",
        "tags": [
          "resources"
        ]
      }
    },
    "/resources/{resourceId}/status": {
      "post": {
        "operationId": "ResourcesController_updateResourceStatus",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "description": "Resource UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourcePublicStatusDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Status updated"
          },
          "400": {
            "description": "Invalid status transition or UUID"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Not authorized to update this resource status"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the operational public status of a resource (owner or coordinator)",
        "tags": [
          "resources"
        ]
      }
    },
    "/emergencies/{emergencyId}/resources/mine": {
      "get": {
        "operationId": "ResourcesController_listMyResources",
        "parameters": [
          {
            "name": "emergencyId",
            "required": true,
            "in": "path",
            "description": "Emergency UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of own resources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceViewDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List resources owned by the authenticated user in an emergency",
        "tags": [
          "resources"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResponseGrid API",
    "description": "API for humanitarian emergency resource coordination",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "LocationDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "Calle Mayor 1, Valencia"
          },
          "latitude": {
            "type": "number",
            "example": 39.4699,
            "description": "Latitude between -90 and 90"
          },
          "longitude": {
            "type": "number",
            "example": -0.3763,
            "description": "Longitude between -180 and 180"
          }
        },
        "required": [
          "address",
          "latitude",
          "longitude"
        ]
      },
      "SupplyLineDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Water bottles",
            "description": "Name of the supply"
          },
          "quantity": {
            "type": "number",
            "example": 100,
            "description": "Quantity (positive integer)"
          },
          "unit": {
            "type": "string",
            "example": "liters",
            "description": "Unit of measurement (optional)"
          },
          "category": {
            "type": "string",
            "enum": [
              "food",
              "water",
              "hygiene",
              "clothing",
              "medical",
              "shelter",
              "tools",
              "other",
              "medicines",
              "medical_equipment",
              "medical_supplies",
              "medical_personnel"
            ],
            "example": "water"
          },
          "presentation": {
            "type": "string",
            "example": "ampolla",
            "description": "Presentation / route of administration: ampolla, EV (intravenoso), inhalador, pastilla, jarabe… Optional, free-form (#61)."
          }
        },
        "required": [
          "name",
          "quantity",
          "category"
        ]
      },
      "RegisterResourceDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "collection_point",
              "delivery_point",
              "collection_and_delivery",
              "warehouse",
              "transport",
              "supplier",
              "venue"
            ],
            "example": "collection_point"
          },
          "stage": {
            "type": "string",
            "enum": [
              "origin",
              "intermediate",
              "destination"
            ],
            "example": "origin",
            "description": "Stage of the resource in the emergency supply chain"
          },
          "name": {
            "type": "string",
            "example": "Cruz Roja Madrid",
            "minLength": 2
          },
          "description": {
            "type": "string",
            "example": "Centro de acopio principal"
          },
          "location": {
            "description": "Physical location of the resource",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocationDto"
              }
            ]
          },
          "ownerOrganizationId": {
            "type": "string",
            "format": "uuid",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "description": "Organization offering this resource (optional)"
          },
          "contact": {
            "type": "string",
            "example": "+58 212 555 0000",
            "description": "Contact info for this resource point"
          },
          "schedule": {
            "type": "string",
            "example": "Lun-Vie 08-18",
            "description": "Operating schedule"
          },
          "manager": {
            "type": "string",
            "example": "Juan Pérez",
            "description": "Responsible manager name"
          },
          "accepts": {
            "example": [
              "water",
              "food"
            ],
            "description": "Category slugs this resource accepts",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "country": {
            "type": "string",
            "example": "VE",
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "city": {
            "type": "string",
            "example": "Caracas",
            "description": "City name"
          },
          "isFinalRecipient": {
            "type": "boolean",
            "example": true,
            "description": "Mark this resource as a final recipient of aid (requires the destination stage)"
          },
          "recipientType": {
            "type": "string",
            "example": "hospital",
            "description": "Recipient type slug (see the emergency recipient-type taxonomy)"
          },
          "items": {
            "description": "Declared inventory: the supply lines this place holds for delivery (optional)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupplyLineDto"
            }
          }
        },
        "required": [
          "type",
          "stage",
          "name",
          "location"
        ]
      },
      "RegisterResourceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          }
        },
        "required": [
          "id"
        ]
      },
      "VerifyResourceDto": {
        "type": "object",
        "properties": {}
      },
      "EditResourceDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Motivo de la edición (obligatorio, para trazabilidad)",
            "minLength": 3,
            "maxLength": 1000,
            "example": "Se corrige el nombre y se completa el horario"
          },
          "name": {
            "type": "string",
            "description": "Nuevo nombre (omitir para no cambiarlo)",
            "minLength": 2
          },
          "description": {
            "type": "string",
            "description": "Nueva descripción. Cadena vacía la borra. Omitir para no cambiarla.",
            "nullable": true
          },
          "contact": {
            "type": "string",
            "description": "Nuevo contacto. Cadena vacía lo borra. Omitir para no cambiarlo.",
            "nullable": true
          },
          "schedule": {
            "type": "string",
            "description": "Nuevo horario. Cadena vacía lo borra. Omitir para no cambiarlo.",
            "nullable": true
          }
        },
        "required": [
          "reason"
        ]
      },
      "DiscardResourceDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Motivo del descarte (obligatorio, para trazabilidad)",
            "minLength": 3,
            "maxLength": 1000,
            "example": "Punto duplicado / fuera del ámbito de la emergencia"
          }
        },
        "required": [
          "reason"
        ]
      },
      "UpdateResourcePublicStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "saturated",
              "paused",
              "closed"
            ],
            "example": "saturated",
            "description": "Target operational status. Hidden is not allowed; use close() to deactivate."
          }
        },
        "required": [
          "status"
        ]
      },
      "LocationViewDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "Calle Mayor 1, Valencia"
          },
          "latitude": {
            "type": "number",
            "example": 39.4699
          },
          "longitude": {
            "type": "number",
            "example": -0.3763
          }
        },
        "required": [
          "address",
          "latitude",
          "longitude"
        ]
      },
      "ResourceViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "type": {
            "type": "string",
            "enum": [
              "collection_point",
              "delivery_point",
              "collection_and_delivery",
              "warehouse",
              "transport",
              "supplier",
              "venue"
            ],
            "example": "collection_point"
          },
          "stage": {
            "type": "string",
            "enum": [
              "origin",
              "intermediate",
              "destination"
            ],
            "example": "origin"
          },
          "name": {
            "type": "string",
            "example": "Cruz Roja Madrid"
          },
          "description": {
            "type": "string",
            "example": "Centro de acopio principal",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/LocationViewDto"
          },
          "verificationLevel": {
            "type": "string",
            "enum": [
              "unverified",
              "verified",
              "official",
              "rejected"
            ],
            "example": "verified"
          },
          "publicStatus": {
            "type": "string",
            "enum": [
              "hidden",
              "active",
              "saturated",
              "paused",
              "closed"
            ],
            "example": "active"
          },
          "ownerOrganizationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "accepts": {
            "example": [
              "water",
              "food"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contact": {
            "type": "string",
            "example": "+58 212 555 0000",
            "nullable": true
          },
          "schedule": {
            "type": "string",
            "example": "Lun-Vie 08-18",
            "nullable": true
          },
          "manager": {
            "type": "string",
            "example": "Juan Pérez",
            "nullable": true
          },
          "sourceName": {
            "type": "string",
            "example": "acopiove.org",
            "nullable": true
          },
          "externalUpdatedAt": {
            "type": "string",
            "example": "2026-06-27T00:00:00.000Z",
            "description": "ISO 8601 date string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "example": "Venezuela",
            "description": "Country string as stored by the ingestion source (e.g. full Spanish name \"Venezuela\"). NOT guaranteed to be an ISO 3166-1 alpha-2 code — value depends on the source `pais` field.",
            "nullable": true
          },
          "city": {
            "type": "string",
            "example": "Caracas",
            "nullable": true
          },
          "isFinalRecipient": {
            "type": "boolean",
            "example": false,
            "description": "Whether this resource is a final recipient of aid"
          },
          "recipientType": {
            "type": "string",
            "example": "hospital",
            "nullable": true,
            "description": "Recipient type slug (see the emergency recipient-type taxonomy)"
          }
        },
        "required": [
          "id",
          "type",
          "stage",
          "name",
          "description",
          "location",
          "verificationLevel",
          "publicStatus",
          "ownerOrganizationId",
          "accepts",
          "contact",
          "schedule",
          "manager",
          "sourceName",
          "externalUpdatedAt",
          "country",
          "city",
          "isFinalRecipient",
          "recipientType"
        ]
      },
...
    }
  }
}